profile page assigns game information to all individual games and displays it to user
