+++ frontend/src/components/AllowedShipDisplayer/AllowedShipDisplayer.js
@@ -6,13 +6,11 @@ const AllowedShipDisplayer = ({ allowedShips }) => {
     return (
         <div className='allowed-ship-container'>
         <h1>Allowed Ships</h1>
-        <ul>
           {Object.entries(allowedShips).map(([ship, quantity]) => (
             <ul key={ship}>
               {ship} - {quantity}
             </ul>
           ))}
-        </ul>
         </div>
     )
 }
