--- frontend/src/pages/Profile/Loss/Loss.js
@@ -1,10 +1,0 @@
-import React from "react";
-import propTypes from "prop-types";
-
-const Loss = ({ game }) => {
-  return <h5>Lost</h5>;
-};
-
-Loss.propTypes = { game: propTypes.object };
-
-export default Loss;
