+++ frontend/src/pages/Profile/Profile.js
@@ -1,11 +1,27 @@
-import React from "react";
-import GamesPlayed from "./GamesPlayed/GamesPlayed";
+import React, { useEffect, useState } from "react";
+import GamesPlayed from "./components/GamesPlayed/GamesPlayed";
+
+import { getHistory } from "../../services/profile";
 
 const Profile = () => {
+  const [gamesHistory, setGamesHistory] = useState([]);
+
+  const getAllGamesOfUser = async () => {
+    const result = await getHistory();
+    setGamesHistory(result);
+  };
+
+  useEffect(() => {
+    getAllGamesOfUser();
+  }, []);
+
+  if (gamesHistory === []) {
+    return;
+  }
+
   return (
     <div>
-      <h1>profile page</h1>
-      <GamesPlayed />
+      <GamesPlayed gamesHistory={gamesHistory} />
     </div>
   );
 };
