--- frontend/src/pages/Profile/Unfinished/Unfinished.js
@@ -1,10 +1,0 @@
-import React from "react";
-import propTypes from "prop-types";
-
-const Unfinished = ({ game }) => {
-  return <h5>Unfinished</h5>;
-};
-
-Unfinished.propTypes = { game: propTypes.object };
-
-export default Unfinished;
