--- frontend/src/pages/Profile/Won/Won.js
@@ -1,15 +1,0 @@
-import React from "react";
-import propTypes from "prop-types";
-
-const Won = ({ game }) => {
-  // const {game_id, }
-    return (
-  <div>
-
-  </div>
-  );
-};
-
-Won.propTypes = { game: propTypes.object };
-
-export default Won;
