+++ frontend/src/pages/Profile/components/GamesPlayed/GamesPlayed.css
@@ -1,0 +1,26 @@
+.container {
+  background-color: #f2f2f2;
+  padding: 10px;
+  margin-bottom: 10px;
+  border-radius: 5px;
+}
+
+.container.unfinished {
+  background-color: #f2f2f2;
+}
+
+.container.loss {
+  background-color: #f36262;
+}
+
+.container.win {
+  background-color: #67f362;
+}
+
+.history-container {
+  display: flex;
+  flex-direction: column;
+  max-width: max-content;
+  height: 85vh;
+  overflow: auto;
+}
