+++ frontend/src/utils/turn.js
@@ -1,0 +1,9 @@
+export const whoseTurnIsIt = (gameState, user_id) => {
+    return (gameState.players.indexOf(user_id) +
+      gameState.turn +
+      gameState.who_started) %
+      2 ===
+      0
+      ? "Your"
+      : "Opponent";
+  };
